# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API SISCAP",
        "version": "1.0"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Guardar nuevos usuarios",
                "parameters": [
                    {
                        "name": "Usuarios",
                        "in": "path",
                        "description": "Datos del nuevo usuario",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guardar nuevos usuarios."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/usuarios/show": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar un usuario",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id del usuario a mostrar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un usuario."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/usuarios/{id}/edit": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener un usuario especifico",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id del usuario a mostrar",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener un usuario especifico."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/usuarios/{usuario}": {
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario especifico",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "description": "Datos del usuario a actualizar",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actualizar un usuario especifico."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/usuarios/{id}": {
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario especifico",
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id del usuario a eliminar",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar un usuario especifico."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/usuarios-roles/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar roles actuales de un usuario",
                "responses": {
                    "200": {
                        "description": "Listar roles actuales de un usuario."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar roles actuales de un usuario",
                "responses": {
                    "200": {
                        "description": "Actualizar roles actuales de un usuario."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/usuario-buscar/{criterio}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Buscar un usuario por nombre",
                "responses": {
                    "200": {
                        "description": "Mostrar usuarios seg\u00fan criterio de busqueda."
                    },
                    "default": {
                        "description": "Falla inesperada. Intente luego"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}